python/agents/product-catalog-ad-generation/pyproject.toml
[project]
name = "content-gen-agent"
version = "0.1.0"
description = "Product Catalog Ad Generation Agent"
authors = [
    { name = "Google LLC" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-adk==1.18.0",
    "google-cloud-aiplatform[adk,agent-engines]==1.126.1",
    "google-cloud-bigquery>=3.35.1",
    "google-genai==1.49.0",
    "langchain>=0.3.27",
    "pandas>=2.3.1",
    "toolbox-core>=0.4.0",
    "moviepy>=2.2.1",
    "google-cloud-texttospeech>=2.29.0",
    "soundfile>=0.13.1",
    "aiohttp>=3.12.15",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.7",
    "python-dotenv>=1.1.1",
]
deployment = [
    "absl-py>=2.3.1",
]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
ignore = ["PLR0911", "C901"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.hatch.build.targets.wheel]
packages = ["content_gen_agent"]

[tool.ruff.lint.isort]
known-first-party = ["content_gen_agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"